Updating case generation to new algorithm
The algorithm is documented here:
https://docs.google.com/document/d/1cYEeIZKQsFHB9tUCi7Ieb3dCXKx-9S3tFEzo
DTcTq7s/edit

Also modified the way the NPC stores alibis as Dialoguer works a
different way as we had it set up before.

We need to add more weapons and update the NPC prefabs.

There are still a few changes that need to be done